+++ src/ts/VAST/AdUnits/VastAdUnitParametersFactory.ts
@@ -9,6 +9,7 @@ import { VastOpenMeasurementController } from 'Ads/Views/OpenMeasurement/VastOpe
 import { VastAdVerification } from 'VAST/Models/VastAdVerification';
 import { OpenMeasurementAdViewBuilder } from 'Ads/Views/OpenMeasurement/OpenMeasurementAdViewBuilder';
 import { ThirdPartyEventMacro } from 'Ads/Managers/ThirdPartyEventManager';
+import { ProgrammaticTrackingService, OMMetric } from 'Ads/Utilities/ProgrammaticTrackingService';
 import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
 
 export class VastAdUnitParametersFactory extends AbstractAdUnitParametersFactory<VastCampaign, IVastAdUnitParameters> {
@@ -74,8 +75,11 @@ export class VastAdUnitParametersFactory extends AbstractAdUnitParametersFactory
             vastAdUnitParameters.om = omManager;
 
             // For brandv1 and brandv2 tracking
-            baseParams.thirdPartyEventManager.setTemplateValue(ThirdPartyEventMacro.OM_ENABLED, `${baseParams.campaign.isOMEnabled()}`);
+            baseParams.thirdPartyEventManager.setTemplateValue(ThirdPartyEventMacro.OM_ENABLED, 'true');
             baseParams.thirdPartyEventManager.setTemplateValue(ThirdPartyEventMacro.OM_VENDORS, omVendors.join('|'));
+            if (baseParams.campaign.getSeatId() === 9078) {
+                ProgrammaticTrackingService.reportMetricEvent(OMMetric.OMEnabledLiftOff);
+            }
         }
 
         return vastAdUnitParameters;